User service implementation (#4)
Setup the "hexagonal" architecture for the user micro service
with which layer separated in a way that technology changing won't
affect the user service logic. Implement the initial part of the user
service: be able to login a user with a basic implementation of the user
service, where the user is ensured to have passed the correct
password, and not flagged as disabled, returning tokens for the
web usage.